Organize Providers into sections
