+++ src/Providers/Validators/DefaultIssueValidator.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using clio.Model;
 
-namespace clio.Providers
+namespace clio.Providers.Validators
 {
 	public class DefaultIssueValidator : BaseIssueValidator
 	{
