+++ src/Providers/Validators/VstsIssueValidator.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Threading.Tasks;
 using clio.Model;
 
-namespace clio.Providers
+namespace clio.Providers.Validators
 {
 	public class VstsIssueValidator : BaseIssueValidator
 	{
